Removed CheckDistance trait - no longer required, within_distance checks moved to compatibility module.
