+++ src/rust/compatibility/python.rs
@@ -1,5 +1,15 @@
-// Python implementation to add PyMethods to CalculationMethod enum struct.
-// but uses ndarrays for all parameters and returns.
+/// Python compatibility layer with `CalculationMethod`.
+///
+/// This module contains the actual `#[pymethods]` called by Python. Instead of taking
+/// `ndarray::ArrayBase` for parameters and return, this module carries out all the
+/// necessary conversions between Rust `ndarray` and Python `numpy` arrays.
+/// i.e. the methods of this module should typically accepts `&PyArray` and returns
+/// PyResult<PyObject>.
+///
+/// It also carries out any checks for concepts that do not exist in Rust, for instance
+/// identity checks between input arrays `s` and `e`, and choose the correct array
+/// accordingly.
+
 use pyo3::prelude::*;
 use pyo3::types::{
     PyTuple,
